requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
numpy
groq
python-dotenv
